refactor: convert component type declarations into a single file
