--- packages/interface/src/common/index.ts
@@ -1,3 +1,0 @@
-export * from "./icon";
-export * from "./image";
-export * from "./typography";
