--- packages/interface/src/data-display/index.ts
@@ -1,7 +1,0 @@
-export * from "./badge";
-export * from "./chip";
-export * from "./divider";
-export * from "./image-avatar";
-export * from "./letter-avatar";
-export * from "./stats";
-export * from "./tag";
