--- packages/interface/src/surface/index.ts
@@ -1,4 +1,0 @@
-export * from "./carousel";
-export * from "./collapse";
-export * from "./grid-list";
-export * from "./stacked-list";
